node: move p2p_node from the network module to the node module
The v1.0 coroutine rewrite (#151) put the whole P2P orchestration — the accept
loop, the peer supervisor, and the message dispatch / serving — inside the
network module. But serving messages needs blockchain access (BIP-35
mempool -> inv, getdata for blocks/txs), and only the node module depends on
both network and blockchain; the network module cannot (and should not) depend
on blockchain.

Move p2p_node, the message dispatcher / message_result, and the ping/pong
handlers to the node module (namespace kth::network -> kth::node). The
connection-level primitives (peer_session, peer_manager, protocols_coro,
peer_database) stay in network, which p2p_node uses via `using namespace
kth::network`. This mirrors the net / net_processing split in BCHN/Core and lets
the node bridge networking to the chain.

Pure relocation, no behavior change. network + node + node-exe + c-api build;
node tests pass (158 assertions).
