+++ src/node/src/handlers/ping.cpp
@@ -2,12 +2,13 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <kth/network/handlers/ping.hpp>
+#include <kth/node/handlers/ping.hpp>
 
 #include <kth/domain.hpp>
-#include <kth/network/p2p_node.hpp>
+#include <kth/node/p2p_node.hpp>
 
-namespace kth::network::handlers::ping {
+namespace kth::node::handlers::ping {
+using namespace kth::network;
 
 ::asio::awaitable<message_result> handle(
     peer_session& peer,
@@ -26,4 +27,4 @@ namespace kth::network::handlers::ping {
     co_return message_result::handled;
 }
 
-} // namespace kth::network::handlers::ping
+} // namespace kth::node::handlers::ping
